python/motap/requirements.txt
pyyaml==5.3.1
fastapi==0.63.0
uvicorn==0.13.3
fuzzymatcher==0.0.5
fuzzywuzzy==0.18.0